test(object_detail): make layout-pack tests pack-discriminating; hoist singleton-reset fixture to conftest
